Start moving metainfo series logic from filter/series to metainfo_series plugin.
